Fix -Wformat warnings.
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@27325 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/core/m_join.c
@@ -282,7 +282,8 @@ m_join(struct Client *client_p, struct Client *source_p, int parc, const char *p
 
 			sendto_one(source_p, form_str(RPL_TOPICWHOTIME),
 				   me.name, source_p->name, chptr->chname,
-				   chptr->topic->topic_info, chptr->topic->topic_time);
+				   chptr->topic->topic_info,
+				   (unsigned long)chptr->topic->topic_time);
 		}
 
 		channel_member_names(chptr, source_p, 1);
